+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## v1 rework with go-chi
 ### Doing
-
+- onboarding checklist
 
 ### Done
-- 2025-01-18 : chore: update go modules and go main version
+- 2025-01-18 : 
+    - chore: update go modules and go main version
+    - NewBranch onboarding checklist
+        - feat: add onboardingCheckList param
 - 2025-01-12 : feat: force new login on client change 
 - 2024-12-23 : 
     - fix: test param and first accounts on user creation